+++ CHANGELOG.md
@@ -2,6 +2,24 @@
 
 All notable changes to the **Hearts of Modding** extension will be documented in this file.
 
+## [0.1.0] - 2026-05-01
+
+### Added
+
+- Initial implementation of CWT schema parsing for triggers and effects.
+
+#### Styling
+
+- Added code action to remove all trailing whitespaces in the file.
+
+- Added "Convert all mixed indentation to tabs in this file" code action.
+
+### Fixed
+
+#### Styling
+
+- Fixed code action "Convert indentation to tabs" not doing anything.
+
 ## [0.0.1] - 2026-04-26
 
 Initial release.